Refactored tag interfaces.
