+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/AuthDomainGroupRoleSpec.scala
@@ -1,7 +1,7 @@
 package org.broadinstitute.dsde.test.api
 
 import cats.implicits.catsSyntaxOptionId
-import org.broadinstitute.dsde.test.api.tagannotation.AuthDomainsTest1
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.AuthDomainsTest1
 import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.AuthDomainsTest
 import org.broadinstitute.dsde.test.util.AuthDomainMatcher
 import org.broadinstitute.dsde.workbench.auth.AuthToken
@@ -15,7 +15,7 @@ import org.broadinstitute.dsde.workbench.service.{Rawls, Sam}
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.must.Matchers
 
-@AuthDomainsTest
+@AuthDomainsTest1
 class AuthDomainGroupRoleSpec
   extends AnyFreeSpec
     with WorkspaceFixtures
