+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/AuthDomainSpec.scala
@@ -1,7 +1,7 @@
 package org.broadinstitute.dsde.test.api
 
 import cats.implicits.catsSyntaxOptionId
-import org.broadinstitute.dsde.test.api.tagannotation.AuthDomainsTest1
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.AuthDomainsTest1
 import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.AuthDomainsTest
 import org.broadinstitute.dsde.workbench.auth.AuthTokenScopes.billingScopes
 import org.broadinstitute.dsde.workbench.config.{ServiceTestConfig, UserPool}
@@ -14,7 +14,7 @@ import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.time.{Seconds, Span}
 
-@AuthDomainsTest
+@AuthDomainsTest1
 class AuthDomainSpec
   extends AnyFlatSpec
     with Matchers
