+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/BillingApiSpec.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.dsde.test.api
 
 import akka.http.scaladsl.model.headers.OAuth2BearerToken
 import com.typesafe.scalalogging.LazyLogging
-import org.broadinstitute.dsde.test.api.tagannotation.BillingsTest1
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.BillingsTest1
 import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.BillingsTest
 import org.broadinstitute.dsde.workbench.auth.{AuthToken, AuthTokenScopes}
 import org.broadinstitute.dsde.workbench.config.{Credentials, ServiceTestConfig, UserPool}
@@ -24,7 +24,7 @@ import scala.concurrent.duration.DurationLong
 import scala.util.Try
 
 //noinspection NoTailRecursionAnnotation,RedundantBlock,ScalaUnusedSymbol
-@BillingsTest
+@BillingsTest1
 class BillingApiSpec
   extends AnyFreeSpec
     with MethodFixtures
