+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/MethodConfigApiSpec.scala
@@ -1,7 +1,7 @@
 package org.broadinstitute.dsde.test.api
 
 import com.typesafe.scalalogging.LazyLogging
-import org.broadinstitute.dsde.test.api.tagannotation.MethodsTest1
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.MethodsTest1
 import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.MethodsTest
 import org.broadinstitute.dsde.workbench.auth.AuthToken
 import org.broadinstitute.dsde.workbench.auth.AuthTokenScopes.billingScopes
