+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/RawlsApiSpec.scala
@@ -5,7 +5,7 @@ import akka.testkit.TestKit
 import cats.implicits.catsSyntaxOptionId
 import com.fasterxml.jackson.databind.{JsonNode, ObjectMapper}
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
-import org.broadinstitute.dsde.test.api.tagannotation.rawls.{MethodsTestObj, RawlsTestObj, WorkspacesTestObj}
+import org.broadinstitute.dsde.test.api.tagannotation.rawls.{MethodsTestObj1, RawlsTestObj1, WorkspacesTestObj1}
 import org.broadinstitute.dsde.workbench.auth.AuthToken
 import org.broadinstitute.dsde.workbench.auth.AuthTokenScopes.billingScopes
 import org.broadinstitute.dsde.workbench.config.{Credentials, ServiceTestConfig, UserPool}
@@ -246,7 +246,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should be able to create workspace and run sub-workflow tasks in non-US regions" taggedAs(RawlsTestObj) in {
+    "should be able to create workspace and run sub-workflow tasks in non-US regions" taggedAs(RawlsTestObj1) in {
       implicit val token: AuthToken = studentBToken
 
       // this will create a method with a workflow containing 3 sub-workflows
@@ -356,7 +356,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should be able to run sub-workflow tasks in a cloned workspace in non-US regions" taggedAs(RawlsTestObj) in {
+    "should be able to run sub-workflow tasks in a cloned workspace in non-US regions" taggedAs(RawlsTestObj1) in {
       implicit val token: AuthToken = studentBToken
 
       // this will create a method with a workflow containing 3 sub-workflows
@@ -472,7 +472,7 @@ class RawlsApiSpec
 
 //    Disabling this test until we decide what to do with it. See AP-177
 
-    "should retrieve metadata with widely scattered sub-workflows in a short time" taggedAs(MethodsTestObj) ignore {
+    "should retrieve metadata with widely scattered sub-workflows in a short time" taggedAs(MethodsTestObj1) ignore {
       implicit val token: AuthToken = studentAToken
 
       val scatterWidth = 500
@@ -566,7 +566,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should label low security bucket" taggedAs(WorkspacesTestObj) in {
+    "should label low security bucket" taggedAs(WorkspacesTestObj1) in {
       implicit val patienceConfig: PatienceConfig = PatienceConfig(timeout = 20 seconds)
       implicit val token: AuthToken = studentAToken
 
@@ -580,7 +580,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should label high security bucket" taggedAs(WorkspacesTestObj) in {
+    "should label high security bucket" taggedAs(WorkspacesTestObj1) in {
       implicit val patienceConfig: PatienceConfig = PatienceConfig(timeout = 20 seconds)
       implicit val token: AuthToken = studentAToken
 
@@ -604,7 +604,7 @@ class RawlsApiSpec
     // We will check against just the suffix instead of the entire string
     val policyToBucketRole = Map("project-owner" -> "terraBucketWriter", "owner" -> "terraBucketWriter", "writer" -> "terraBucketWriter", "reader" -> "terraBucketReader")
 
-    "should have correct policies in Sam and IAM roles in Google when an unconstrained workspace is created" taggedAs(WorkspacesTestObj) in {
+    "should have correct policies in Sam and IAM roles in Google when an unconstrained workspace is created" taggedAs(WorkspacesTestObj1) in {
       implicit val patienceConfig: PatienceConfig = PatienceConfig(timeout = 20 seconds)
       implicit val token: AuthToken = ownerAuthToken
 
@@ -626,7 +626,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should have correct policies in Sam and IAM roles in Google when a constrained workspace is created" taggedAs(WorkspacesTestObj) in {
+    "should have correct policies in Sam and IAM roles in Google when a constrained workspace is created" taggedAs(WorkspacesTestObj1) in {
       implicit val patienceConfig: PatienceConfig = PatienceConfig(timeout = 20 seconds)
       implicit val token: AuthToken = ownerAuthToken
 
@@ -648,7 +648,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should clone a workspace and only copy files in the specified path" taggedAs(WorkspacesTestObj) in {
+    "should clone a workspace and only copy files in the specified path" taggedAs(WorkspacesTestObj1) in {
       implicit val patienceConfig: PatienceConfig = PatienceConfig(timeout = 5 minutes)
       implicit val token: AuthToken = studentAToken
 
@@ -688,7 +688,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should support running workflows with private docker images" taggedAs(MethodsTestObj) in {
+    "should support running workflows with private docker images" taggedAs(MethodsTestObj1) in {
       implicit val token: AuthToken = ownerAuthToken
 
       val privateMethod: Method = MethodData.SimpleMethod.copy(
@@ -751,7 +751,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should support running workflows with wdl structs" taggedAs(MethodsTestObj) in {
+    "should support running workflows with wdl structs" taggedAs(MethodsTestObj1) in {
       implicit val token: AuthToken = ownerAuthToken
 
       val privateMethod: Method = MethodData.SimpleMethod.copy(
@@ -924,7 +924,7 @@ class RawlsApiSpec
       }(owner.makeAuthToken(billingScopes))
     }
 
-    "should fail to launch a submission with a reserved output attribute" taggedAs(MethodsTestObj) in {
+    "should fail to launch a submission with a reserved output attribute" taggedAs(MethodsTestObj1) in {
       implicit val token: AuthToken = ownerAuthToken
 
       withTemporaryBillingProject(billingAccountId) { projectName =>
