--- automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/MethodsTest1.java
@@ -1,9 +1,0 @@
-package org.broadinstitute.dsde.test.api.tagannotation;
-
-import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestAnnotationBase;
-import org.scalatest.TagAnnotation;
-
-@TagAnnotation
-@RawlsTestAnnotationBase
-public @interface MethodsTest1 {
-}
