--- automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/WorkspacesTest1.java
@@ -1,8 +1,0 @@
-package org.broadinstitute.dsde.test.api.tagannotation;
-
-import org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestAnnotationBase;
-import org.scalatest.TagAnnotation;
-
-@TagAnnotation
-@RawlsTestAnnotationBase
-public @interface WorkspacesTest1 {}
