+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/rawls/RawlsTestGroupAnnotations.java
@@ -7,7 +7,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-public interface RawlsTestGroupAnnotations {
+public class RawlsTestGroupAnnotations {
     @TagAnnotation
     @Retention(RetentionPolicy.RUNTIME)
     @Target({ElementType.METHOD, ElementType.TYPE})
