+++ automation/src/test/scala/org/broadinstitute/dsde/test/api/tagannotation/rawls/RawlsTestTaggedAs.scala
@@ -2,6 +2,9 @@ package org.broadinstitute.dsde.test.api.tagannotation.rawls
 
 import org.scalatest.Tag
 
+object MethodsTestObj1 extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.MethodsTest1")
+object RawlsTestObj1 extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTest1")
+object WorkspacesTestObj1 extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.WorkspacesTest1")
 object MethodsTestObj extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.MethodsTest")
 object RawlsTestObj extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.RawlsTest")
 object WorkspacesTestObj extends Tag("org.broadinstitute.dsde.test.api.tagannotation.rawls.RawlsTestGroupAnnotations.WorkspacesTest")
